requirements.txt
httpx~=0.28.1
pydantic~=2.11.7
pydantic-settings~=2.10.1
allure-python-commons~=2.14.3
pytest~=8.4.1
allure-pytest~=2.14.3
mimesis~=18.0.0
jsonschema~=4.25.0
pytest-playwright
